refactor: extract GeocoderEngine from DataProcessor (Phase 2c stage 2)
The address/place candidate search moves verbatim into
Services/GeocoderEngine.cs, backed by the shared DuckDBManager.
DataProcessor keeps thin delegating wrappers so OvertureGeocoderService
is unchanged. EscapeSqlLiteral is promoted to GeoParquetSql (its only
users were the geocoder queries) with unit test coverage.
